competitors_scraper/nttdeta/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
openai==1.3.7
python-dotenv==1.0.0
tqdm==4.66.1
